control_panel_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[hiredis]==5.0.7
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-multipart==0.0.9
email-validator==2.2.0
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.24.0
